requirements.txt
sqlalchemy
python-dotenv
fastapi
uuid
typing
passlib
pydantic
requests
psycopg2-binary